mobile popup handle test code
