+++ src/test/java/com/mobile/tests/MobileTestBase.java
@@ -56,7 +56,10 @@ public abstract class MobileTestBase {
 
 
     @AfterMethod(alwaysRun = true)
-    public void tearDown() {
-        if (driver != null) driver.quit();
+    public void tearDown() throws InterruptedException {
+        if (driver != null) {
+            Thread.sleep(2000);
+            driver.quit();
+        }
     }
 }
